requirements.txt
streamlit
streamlit-navigation-bar
pandas
numpy
matplotlib
scikit-learn
tensorflow
pillow